Made the new version FreePascal clean.  I decoupled the requirement to use dwe by making pcre.pas and regexp.pas apart of this source tree.
Apparently, freepascal does not clear data on new records, so I have to
explicitly set two booleans to false every time.  Plus the install routine
for fpc is awkward.  I now remember why I prefer the old, outdated kylix.
I'm still not a fan of fpc, but here it is.